LDTOOLS-142: MSA Viewer: Improve selection behavior and provide APIs to get and set selection data (#22)
Implemented Ctrl/Cmd selection, Ctrl/Cmd de-selection, Shift selection, and complete de-selection when clicking on the viewer's white space.

Introduced two new APIs: getSelectionData() and setSelectionData().
getSelectionData() returns the selection data (maintained in terms of models) as an object.
setSelectionData() can be used to reset the selection in MSA by supplying selection data as an object.

The format of the selection data is as follows:
```
selectionData = {
  selectedRows: (string | number)[]
  selectedColumns: number[]
  selectedPositions: {seqId: string | number; columnIndex: number}[]
  selectedLabels: (string | number)[]
}
```

Co-authored-by: Ritik Garg <ritik.garg@schrodinger.com>